lf polish: tests: consolidate test files to match new package structure
Merged test_backend.py → test_launcher.py, test_automode.py → test_config.py, test_pipeline.py → test_config.py, test_task_args.py → test_context.py. Removed mock-wiring tests from test_collector.py per STYLE.md guidance.
