+++ .design/refactor-vision.md
@@ -2,7 +2,7 @@
 
 ## Status
 
-**Implemented.** Tests pass (387/387).
+**Implemented.** Tests pass (381/381).
 
 ## New Structure
 
@@ -49,6 +49,18 @@ src/loopflow/
 6. Updated all imports (src + tests)
 7. Fixed circular import (lfd → lf.context → lfd.work)
 8. Updated pyproject.toml entrypoints
+
+## Test Consolidation
+
+Refactored test structure to match the new package layout:
+
+1. Merged `test_backend.py` → `test_launcher.py` (runner classes + command building)
+2. Merged `test_automode.py` → `test_config.py` (interactive mode config)
+3. Merged `test_pipeline.py` → `test_config.py` (PipelineConfig tests)
+4. Merged `test_task_args.py` → `test_context.py` (template substitution tests)
+5. Removed mock-wiring tests from `test_collector.py` (STYLE.md: "Mock side effects, but don't test mock wiring")
+
+Result: 22 test files, 381 tests (was 25 files, 387 tests).
 
 ## Open Questions
