+++ tests/test_launcher.py
@@ -1,6 +1,9 @@
-"""Tests for loopflow.launcher module."""
+"""Tests for loopflow.lf.launcher module."""
 
 from pathlib import Path
+from unittest.mock import patch, MagicMock
+
+import pytest
 
 from loopflow.lf.launcher import (
     build_claude_command,
@@ -10,6 +13,11 @@ from loopflow.lf.launcher import (
     build_model_interactive_command,
     check_claude_available,
     _format_normalized_event,
+    get_runner,
+    ClaudeRunner,
+    CodexRunner,
+    GeminiRunner,
+    LaunchResult,
 )
 
 
@@ -389,3 +397,140 @@ def test_format_normalized_event_unknown_type():
     result = _format_normalized_event(event)
 
     assert result is None
+
+
+# =============================================================================
+# Runner abstraction tests
+# =============================================================================
+
+
+def test_get_runner_returns_claude():
+    """get_runner returns ClaudeRunner instance for 'claude'."""
+    runner = get_runner("claude")
+    assert isinstance(runner, ClaudeRunner)
+
+
+def test_get_runner_returns_codex():
+    """get_runner returns CodexRunner instance for 'codex'."""
+    runner = get_runner("codex")
+    assert isinstance(runner, CodexRunner)
+
+
+def test_get_runner_returns_gemini():
+    """get_runner returns GeminiRunner instance for 'gemini'."""
+    runner = get_runner("gemini")
+    assert isinstance(runner, GeminiRunner)
+
+
+def test_get_runner_raises_on_unknown():
+    """get_runner raises ValueError for unknown model name."""
+    with pytest.raises(ValueError, match="Unknown model"):
+        get_runner("gpt4")
+
+
+def test_launch_result_dataclass():
+    """LaunchResult holds exit code and optional output."""
+    result = LaunchResult(exit_code=0, output="success")
+    assert result.exit_code == 0
+    assert result.output == "success"
+
+    result_no_output = LaunchResult(exit_code=1)
+    assert result_no_output.exit_code == 1
+    assert result_no_output.output is None
+
+
+def test_claude_runner_is_available():
+    """ClaudeRunner.is_available() checks for claude CLI."""
+    runner = ClaudeRunner()
+
+    with patch("loopflow.lf.launcher.check_claude_available") as mock_check:
+        mock_check.return_value = True
+        assert runner.is_available() is True
+
+        mock_check.return_value = False
+        assert runner.is_available() is False
+
+
+def test_codex_runner_is_available():
+    """CodexRunner.is_available() checks for codex CLI."""
+    runner = CodexRunner()
+
+    with patch("subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(returncode=0)
+        assert runner.is_available() is True
+
+        mock_run.side_effect = FileNotFoundError()
+        assert runner.is_available() is False
+
+
+def test_claude_runner_launch_returns_result():
+    """ClaudeRunner.launch() returns LaunchResult."""
+    runner = ClaudeRunner()
+
+    with patch("loopflow.lf.launcher.launch_claude") as mock_launch:
+        mock_launch.return_value = (0, "output")
+        result = runner.launch("test prompt")
+
+        assert isinstance(result, LaunchResult)
+        assert result.exit_code == 0
+        assert result.output == "output"
+
+
+def test_codex_runner_launch_auto_mode():
+    """CodexRunner.launch() with auto captures output."""
+    runner = CodexRunner()
+
+    with patch("subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(returncode=0, stdout="codex output")
+        result = runner.launch("test prompt", auto=True)
+
+        assert result.exit_code == 0
+        assert result.output == "codex output"
+
+
+def test_codex_runner_launch_interactive():
+    """CodexRunner.launch() without auto runs interactively."""
+    runner = CodexRunner()
+
+    with patch("subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(returncode=0)
+        result = runner.launch("test prompt", auto=False)
+
+        assert result.exit_code == 0
+        assert result.output is None
+
+
+def test_gemini_runner_is_available():
+    """GeminiRunner.is_available() checks for gemini CLI."""
+    runner = GeminiRunner()
+
+    with patch("loopflow.lf.launcher.check_gemini_available") as mock_check:
+        mock_check.return_value = True
+        assert runner.is_available() is True
+
+        mock_check.return_value = False
+        assert runner.is_available() is False
+
+
+def test_gemini_runner_launch_auto_mode():
+    """GeminiRunner.launch() with auto captures output."""
+    runner = GeminiRunner()
+
+    with patch("subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(returncode=0, stdout="gemini output")
+        result = runner.launch("test prompt", auto=True)
+
+        assert result.exit_code == 0
+        assert result.output == "gemini output"
+
+
+def test_gemini_runner_launch_interactive():
+    """GeminiRunner.launch() without auto runs interactively."""
+    runner = GeminiRunner()
+
+    with patch("subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(returncode=0)
+        result = runner.launch("test prompt", auto=False)
+
+        assert result.exit_code == 0
+        assert result.output is None
